CORE-5530 - Encrypt and decrypt unauthenticated messages (#2313)
Adding ecies encryption/decryption for submitting registration request. Also includes a fix for a minor crypto bug: There was a crypto bug in CipherSchemeMetadataProvider: when we were parsing a public key to PEM string and then converting it back to PublicKey the original algorithm of the key got overridden by the value of 'ECDSA'.